* commit 'e6f3d4268305d2eac76af6115d80f1d2c4919f56 [formerly 6717f02c62df6076da9e901f2514aa9ba24f253c]': (148 commits)
  DF-10716 fix one more time cmake build for win32
  DF-10716 fix asset-server build
  DF-10716 fix cmake dava-template var
  DF-10716 try fix mounting for dependent packs (for Denis, need test. TODO)
  DF-10716 second fix win32 cmake config
  DF-10716 fix cmake config for mac or win
  DF-10716 try fix TestBed package for wint32 (add scripts dir to archive)
  DF-10716 fix UAP build
  DF-10716 fix crush in CurDownloader uint32 -> long 8 bytes on mac os
  DF-10716 fix test bed local server start stop with Terminal and quit from terminal
  DF-10716 test bed works with local python http server (tested on win32)
  DF-10716 fix unit test build
  DF-10716 start fixing unit tests
  DF-10716 fix test_bed pack_manager downloading and building packs
  DF-10716 continue fixing test_bed (not change pack_mananger)
  DF-10716 fix build with trunk client
  DF-10716 change url to server start fixing test bed for new pack generation
  DF-10716 fix not UNITY build after merge
  DF-10716 fix comments review
  [DF-10716] reverting back %<group> usage in commands due to random crashes
  ...


Former-commit-id: 84dc9185f6770cefe3709c9f442b3314581e6b97

+++ Projects/UnitTests/Classes/Tests/FileSystemTest.cpp
@@ -19,7 +19,7 @@ DAVA_TESTCLASS (FileSystemTest)
         TEST_VERIFY(fileList->GetDirectoryCount() == 4);
         TEST_VERIFY(fileList->GetFileCount() == 0);
 
-        for (int32 ifo = 0; ifo < fileList->GetCount(); ++ifo)
+        for (uint32 ifo = 0; ifo < fileList->GetCount(); ++ifo)
         {
             if (fileList->IsNavigationDirectory(ifo))
                 continue;
@@ -34,7 +34,7 @@ DAVA_TESTCLASS (FileSystemTest)
                 TEST_VERIFY(pathname == "~res:/TestData/FileSystemTest/Folder1/");
                 TEST_VERIFY(files->GetFileCount() == 3);
 
-                for (int32 ifi = 0; ifi < files->GetCount(); ++ifi)
+                for (uint32 ifi = 0; ifi < files->GetCount(); ++ifi)
                 {
                     if (files->IsNavigationDirectory(ifi))
                         continue;
@@ -84,7 +84,7 @@ DAVA_TESTCLASS (FileSystemTest)
         TEST_VERIFY(fileList->GetDirectoryCount() == 4);
         TEST_VERIFY(fileList->GetFileCount() == 0);
 
-        for (int32 ifo = 0; ifo < fileList->GetCount(); ++ifo)
+        for (uint32 ifo = 0; ifo < fileList->GetCount(); ++ifo)
         {
             if (fileList->IsNavigationDirectory(ifo))
                 continue;
@@ -99,7 +99,7 @@ DAVA_TESTCLASS (FileSystemTest)
                 TEST_VERIFY(pathname == "~doc:/TestData/FileSystemTest/Folder1/");
                 TEST_VERIFY(files->GetFileCount() == 3);
 
-                for (int32 ifi = 0; ifi < files->GetCount(); ++ifi)
+                for (uint32 ifi = 0; ifi < files->GetCount(); ++ifi)
                 {
                     if (files->IsNavigationDirectory(ifi))
                         continue;
@@ -233,16 +233,16 @@ DAVA_TESTCLASS (FileSystemTest)
         if (!f2 || !f2)
             return;
 
-        uint32 size = f1->GetSize();
+        uint64 size = f1->GetSize();
         TEST_VERIFY(size == f2->GetSize());
 
-        char8* buf1 = new char8[size];
-        char8* buf2 = new char8[size];
+        char8* buf1 = new char8[static_cast<size_t>(size)];
+        char8* buf2 = new char8[static_cast<size_t>(size)];
 
         do
         {
-            uint32 res1 = f1->ReadLine(buf1, size);
-            uint32 res2 = f2->ReadLine(buf2, size);
+            uint32 res1 = f1->ReadLine(buf1, static_cast<uint32>(size));
+            uint32 res2 = f2->ReadLine(buf2, static_cast<uint32>(size));
             TEST_VERIFY(res1 == res2);
             TEST_VERIFY(!Memcmp(buf1, buf2, res1));
 
@@ -259,12 +259,12 @@ DAVA_TESTCLASS (FileSystemTest)
         f1 = File::Create(fileInAssets, File::OPEN | File::READ);
         f2 = File::Create(copyTo, File::OPEN | File::READ);
 
-        int32 seekPos = f1->GetSize() + 10;
+        int64 seekPos = f1->GetSize() + 10;
         TEST_VERIFY(f1->Seek(seekPos, File::SEEK_FROM_START));
         TEST_VERIFY(f2->Seek(seekPos, File::SEEK_FROM_START));
 
-        uint32 pos1 = f1->GetPos();
-        uint32 pos2 = f2->GetPos();
+        uint64 pos1 = f1->GetPos();
+        uint64 pos2 = f2->GetPos();
 
         TEST_VERIFY(f1->IsEof() == false);
         TEST_VERIFY(f2->IsEof() == false);
